refactor: move `pipxInstall` function to `src/action.mts` file
