+++ src/action.mts
@@ -1,0 +1,10 @@
+import core from "@actions/core";
+import { installPackage } from "./pipx/install.mjs";
+
+export async function pipxInstall(...pkgs: string[]): Promise<void> {
+  for (const pkg of pkgs) {
+    await core.group(`Installing \u001b[34m${pkg}\u001b[39m...`, async () => {
+      await installPackage(pkg);
+    });
+  }
+}
